# Server/api/openapi.yaml
...
openapi: "3.1.0"
info:
  title: "DictaDuo API"
  version: "2.0.0"
  description: "Shared transport contract for the Swift client and TypeScript server. Optional properties are omitted rather than encoded as null. Dates use whole-second ISO8601 UTC timestamps."
servers:
  -
    url: "http://localhost:8391"
security:
  -
    bearerAuth: []
...
paths:
  /v1/button-destinations:
    get:
      operationId: getButtonDestinations
      parameters: []
      responses:
        "200":
          description: Live button-destination state. Selection and commands never survive
            expiry or restart.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ButtonDestinationState"
        default:
          $ref: "#/components/responses/APIError"
    post:
      operationId: registerButtonDestination
      parameters:
        - &a1
          name: X-DictaDuo-Destination-Owner
          in: header
          required: true
          schema:
            type: string
            pattern: ^[0-9a-f]{64}$
          description: Fresh 256-bit secret for this in-memory destination registration.
            Never persist or log it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterButtonDestination"
      responses:
        "200":
          description: Live button-destination state. Selection and commands never survive
            expiry or restart.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ButtonDestinationState"
        default:
          $ref: "#/components/responses/APIError"
  /v1/button-destinations/{id}/heartbeat:
    post:
      operationId: heartbeatButtonDestination
      parameters:
        - &a2
          name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - *a1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeartbeatButtonDestination"
      responses:
        "200":
          description: Live button-destination state. Selection and commands never survive
            expiry or restart.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ButtonDestinationState"
        default:
          $ref: "#/components/responses/APIError"
  /v1/button-destinations/{id}/select:
    post:
      operationId: selectButtonDestination
      parameters:
        - *a2
        - *a1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectButtonDestination"
      responses:
        "200":
          description: Live button-destination state. Selection and commands never survive
            expiry or restart.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ButtonDestinationState"
        default:
          $ref: "#/components/responses/APIError"
  /v1/button-destinations/{id}/complete:
    post:
      operationId: completeButtonTake
      parameters:
        - *a2
        - *a1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteButtonTake"
      responses:
        "200":
          description: Live button-destination state. Selection and commands never survive
            expiry or restart.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ButtonDestinationState"
        default:
          $ref: "#/components/responses/APIError"
  /v1/button-destinations/{id}:
    delete:
      operationId: unregisterButtonDestination
      parameters:
        - *a2
        - *a1
      responses:
        "200":
          description: Live button-destination state. Selection and commands never survive
            expiry or restart.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ButtonDestinationState"
        default:
          $ref: "#/components/responses/APIError"
  "/v1/audio-sources":
    get:
      operationId: "listAudioSources"
      responses:
        "200":
          description: "Bounded snapshot; discovery never starts audio or connects Bluetooth.
            Other computers only see shared sources."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/AudioSourceList"
        default:
          "$ref": "#/components/responses/APIError"
  "/v1/audio-sources/sharing":
    put:
      operationId: "setAudioSourceSharing"
      description: "Shares or stops sharing one source with other computers. Only a client on
        the computer that hosts the sources may change this."
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/AudioSourceSharing"
      responses:
        "200":
          description: "Updated snapshot."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/AudioSourceList"
        default:
          "$ref": "#/components/responses/APIError"
  /v1/button-destinations/target:
    put:
      operationId: setButtonTarget
      description: Chooses where the DJI button types. Stored on the server and shared by
        every destination.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ButtonTarget"
      responses:
        "200":
          description: Live button-destination state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ButtonDestinationState"
        default:
          $ref: "#/components/responses/APIError"
  "/v2/captures":
    post:
      operationId: "startCapture"
      parameters:
        - "$ref": "#/components/parameters/CaptureOwner"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/StartCaptureRequest"
      responses:
        "201":
          description: "Admitted recording session with acknowledged microphone readiness."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/RecordingSnapshot"
        default:
          "$ref": "#/components/responses/APIError"
...
  "/v1/preferences":
    get:
      operationId: "getPreferences"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/PreferencesSnapshot"
        default:
          "$ref": "#/components/responses/APIError"
    put:
      operationId: "updatePreferences"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/PreferencesSnapshot"
        default:
          "$ref": "#/components/responses/APIError"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/PreferencesSnapshot"
        x-maximumBytes: 262144
  "/v1/generations":
    get:
      operationId: "listGenerations"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/GenerationPage"
        default:
          "$ref": "#/components/responses/APIError"
      parameters:
        - "$ref": "#/components/parameters/CaptureView"
        -
          name: "limit"
          in: "query"
          required: false
          schema:
            type: "integer"
            minimum: 1
            maximum: 100
            default: 50
        -
          name: "before"
          in: "query"
          required: false
          schema:
            type: "string"
        -
          name: "source"
          in: "query"
          required: false
          schema:
            type: "string"
            enum:
              - "dictaduo"
              - "wispr-flow"
    post:
      operationId: "createGeneration"
      responses:
        "201":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/GenerationRecord"
        default:
          "$ref": "#/components/responses/APIError"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/CreateGenerationRequest"
        x-maximumBytes: 262144
  "/v1/generations/{id}":
    get:
      operationId: "getGeneration"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/GenerationRecord"
        default:
          "$ref": "#/components/responses/APIError"
      parameters:
        - "$ref": "#/components/parameters/CaptureView"
        -
          name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
    delete:
      operationId: "deleteGeneration"
      responses:
        "204":
          description: "Success"
        default:
          "$ref": "#/components/responses/APIError"
      parameters:
        -
          name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
  "/v1/generations/{id}/audio/{kind}":
    post:
      operationId: "appendAudio"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/AudioChunkReceipt"
        default:
          "$ref": "#/components/responses/APIError"
      parameters:
        -
          name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
        -
          name: "kind"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/AudioKind"
        -
          name: "sequence"
          in: "query"
          required: true
          schema:
            type: "integer"
            minimum: 0
            maximum: 4095
        -
          name: "sampleRate"
          in: "query"
          required: true
          schema:
            type: "integer"
            minimum: 8000
            maximum: 192000
        -
          name: "channels"
          in: "query"
          required: true
          schema:
            type: "integer"
            minimum: 1
            maximum: 8
      requestBody:
        required: true
        content:
          "application/octet-stream":
            schema:
              type: "string"
              format: "binary"
        description: "Raw interleaved little-endian float32 PCM. Chunks must contain whole frames. Repeated sequences must have identical bytes."
        x-maximumBytes: 1048576
  "/v1/generations/{id}/finish":
    post:
      operationId: "finishGeneration"
      responses:
        "202":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/GenerationRecord"
        default:
          "$ref": "#/components/responses/APIError"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/FinishGenerationRequest"
        x-maximumBytes: 262144
      parameters:
        -
          name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
  "/v1/generations/{id}/cancel":
    post:
      operationId: "cancelGeneration"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/GenerationRecord"
        default:
          "$ref": "#/components/responses/APIError"
      parameters:
        - "$ref": "#/components/parameters/CaptureView"
        - "$ref": "#/components/parameters/CaptureMutationOwner"
        -
          name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
  "/v1/generations/{id}/retry":
    post:
      operationId: "retryGeneration"
      responses:
        "202":
          description: "Accepted"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/GenerationRecord"
        default:
          "$ref": "#/components/responses/APIError"
      parameters:
        - "$ref": "#/components/parameters/CaptureView"
        -
          name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
...
  "/v1/generations/{id}/events":
    get:
      operationId: "generationEvents"
      responses:
        "200":
          description: "Success"
          content:
            "application/x-ndjson":
              schema:
                type: "string"
              x-recordSchema:
                "$ref": "#/components/schemas/GenerationRecord"
        default:
          "$ref": "#/components/responses/APIError"
      description: "A bounded NDJSON stream of complete GenerationRecord values, with the latest record repeated as a two-second heartbeat. Ends after a terminal record."
      parameters:
        - "$ref": "#/components/parameters/CaptureView"
        -
          name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
  "/v1/generations/{id}/artifacts/{filename}":
    get:
      operationId: "downloadArtifact"
      responses:
        "200":
          description: "Success"
          content:
            "application/octet-stream":
              schema:
                type: "string"
                format: "binary"
            "audio/wav":
              schema:
                type: "string"
                format: "binary"
            "image/png":
              schema:
                type: "string"
                format: "binary"
            "application/json":
              schema:
                type: "string"
                format: "binary"
            "text/plain":
              schema:
                type: "string"
        default:
          "$ref": "#/components/responses/APIError"
      parameters:
        -
          name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
        -
          name: "filename"
          in: "path"
          required: true
          schema:
            type: "string"
            enum:
              - "inference.wav"
              - "original.wav"
              - "metadata.json"
              - "transcript.txt"
              - "source.json"
              - "source.wav"
              - "opus.json"
              - "screenshot.png"
              - "built-in-audio.bin"
  "/v1/imports/wispr-flow/known":
    post:
      operationId: "knownWisprFlowIDs"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/WisprFlowKnownIDsResponse"
        default:
          "$ref": "#/components/responses/APIError"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/WisprFlowKnownIDsRequest"
        x-maximumBytes: 262144
  "/v1/imports/wispr-flow":
    post:
      operationId: "beginWisprFlowImport"
      responses:
        "201":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/WisprFlowImportSession"
        default:
          "$ref": "#/components/responses/APIError"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/WisprFlowImportRequest"
        x-maximumBytes: 262144
  "/v1/imports/wispr-flow/dictionary":
    put:
      operationId: "archiveWisprFlowDictionary"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/WisprFlowDictionaryArchiveReceipt"
        default:
          "$ref": "#/components/responses/APIError"
      requestBody:
        required: true
        content:
          "application/json":
            schema: {}
        x-maximumBytes: 8388608
        description: "Original Wispr Flow dictionary JSON archive. This does not change the active personal dictionary."
...
  "/v1/imports/wispr-flow/{id}/complete":
    post:
      operationId: "completeWisprFlowImport"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/WisprFlowImportResult"
        default:
          "$ref": "#/components/responses/APIError"
      parameters:
        -
          name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
  "/v1/imports/wispr-flow/{id}":
    delete:
      operationId: "cancelWisprFlowImport"
      responses:
        "204":
          description: "Success"
        default:
          "$ref": "#/components/responses/APIError"
      parameters:
        -
          name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
...
components:
  schemas:
    UUID:
      type: "string"
      format: "uuid"
...
    AudioKind:
      type: "string"
      enum:
        - "inference"
        - "original"
...
    RegisterButtonDestination:
      type: object
      additionalProperties: false
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        device:
          $ref: "#/components/schemas/DeviceIdentity"
      required:
        - id
        - device
...
    ButtonDestinationState:
      type: object
      additionalProperties: false
      properties:
        selected:
          $ref: "#/components/schemas/ButtonDestination"
        destinations:
          type: array
          items:
            $ref: "#/components/schemas/ButtonDestination"
          maxItems: 8
        source:
          $ref: "#/components/schemas/AudioSourceIdentity"
        available:
          type: boolean
        command:
          $ref: "#/components/schemas/ButtonCommand"
        buttonTarget:
          $ref: "#/components/schemas/ButtonTarget"
      required:
        - destinations
        - available
...
    HeartbeatButtonDestination:
      type: object
      additionalProperties: false
      properties:
        acknowledgement:
          $ref: "#/components/schemas/UUID"
      required: []
...
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      description: "Optional on loopback. Required for remote binding. Native clients send no Origin header."
  responses:
    APIError:
      description: "A request, admission, or server error."
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/APIErrorResponse"